fix: resolve type mismatches and configuration issues after dev merge
- Fixed ID type inconsistencies: Changed DTOs to use Long for user IDs (reportedById, assignedToId, commentedById, uploadedById) to match Users entity
- Updated request DTOs and service/controller method signatures for type consistency
- Removed unused imports from TicketController, Ticket, TicketComment, Users, and UserController
- Added PasswordEncoder bean (BCryptPasswordEncoder) to SecurityConfig
- Fixed UserMapper with @Mapping annotations for unmapped target properties
- Updated application.yaml with development configuration values for JWT_SECRET and Google OAuth credentials

Resolves database connection and bean initialization errors after merging dev branch.

+++ backend/src/main/java/com/project/smart_campus_operationhub/dtos/TicketAssignDTO.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class TicketAssignDTO {
 
-    private Integer technicianId; // ID of technician to assign
+    private Long technicianId; // ID of technician to assign
     private String priority; // Optional: update priority when assigning
 }
